requirements.txt
certifi==2018.8.24
chardet==3.0.4
dukpy==0.2.0
echarts-china-provinces-pypkg==0.0.2
echarts-countries-pypkg==0.1.4
future==0.16.0
idna==2.7
itchat==1.3.10
javascripthon==0.10
jieba==0.39
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
macropy3==1.1.0b2
MarkupSafe==1.0
Pillow==5.2.0
pyecharts==0.5.9
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pymongo==3.7.1
pypng==0.0.18
PyQRCode==1.2.1
requests==2.19.1
urllib3==1.23
